[MAP-71] 설정한 반경내 스케줄 목록 조회 기능 추가

+++ src/main/java/com/mapnote/mapnoteserver/domain/user/dto/UserRequest.java
@@ -90,7 +90,7 @@ public class UserRequest {
     private String name;
 
     @Schema(description = "반경", example = "8")
-    private Long boundary;
+    private Double boundary;
 
   }
